requirements.txt
discord.py
python-dotenv
pynacl
black
yt_dlp
gtts